requirements.txt
Flask==3.0.3
pycryptodome==3.20.0
requests==2.31.0
gunicorn==22.0.0